requirement.txt
babelfish==0.5.4
enzyme==0.4.1
guessit==0.10.1
IMDbPY==5.0
Pafy==0.3.72
pytvdbapi==0.5.0
pyxdg==0.25
rottentomatoes==2.1
#subliminal==0.8.0.dev0
https://github.com/Diaoul/subliminal/archive/master.zip
tabulate==0.7.4
yify-sub==0.1.0
